requirements.txt
Pillow>=9.5.0
pillow-heif>=0.11.0
customtkinter>=5.2.0
cffi>=1.15.1
pycparser>=2.21
